refactor(quick-commands): 拆出快捷命令树节点、编辑器、详情卡片、对话框与交互 hook
